src/requirements.txt
polars
pandas
duckdb
flask
streamlit
ydata-profiling
kagglehub
fastapi
uvicorn 
fastapi-cache2